# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusOps Expo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lostitems:
    get:
      operationId: getLostItems
      tags: [lostitems]
      summary: Get lost items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of lost items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LostItem"
    post:
      operationId: createLostItem
      tags: [lostitems]
      summary: Create lost item report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLostItemRequest"
      responses:
        "201":
          description: Lost item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
  /lostitems/{id}:
    get:
      operationId: getLostItemById
      tags: [lostitems]
      summary: Get lost item by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lost item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
    patch:
      operationId: updateLostItemStatus
      tags: [lostitems]
      summary: Update lost item status (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLostItemRequest"
      responses:
        "200":
          description: Updated lost item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
    delete:
      operationId: deleteLostItem
      tags: [lostitems]
      summary: Delete lost item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    LostItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          enum: [lost, found, claimed]
        reportedBy:
          type: string
        reportedByName:
          type: string
        location:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - status
        - reportedBy
    CreateLostItemRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        location:
          type: string
      required:
        - title
    UpdateLostItemRequest:
      type: object
      properties:
        status:
          type: string
          enum: [lost, found, claimed]
        description:
          type: string
...